Add subgroup component
