+++ src/components/group.module.css
@@ -24,7 +24,6 @@
   overflow-wrap: break-word;
 }
 
-
 .groupBtn {
   border-radius: 10px;
   width: 70%;
@@ -42,25 +41,4 @@
   padding: 5px;
   height: auto;
   gap: 10px;
-}
-
-.subgroup {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  border: 3px solid var(--color-1);
-  border-radius: 10px;
-  padding: 5px;
-  height: auto;
-  gap: 10px;
-}
-
-.subGroupBtn {
-  border-radius: 10px;
-  width: 100%;
-  height: 30px;
-  cursor: pointer;
-  background-color: var(--color-1);
-  border: 2px solid var(--color-5);
 }
