+++ src/components/subgroup.module.css
@@ -1,0 +1,23 @@
+.subgroup {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  border: 3px solid var(--color-1);
+  border-radius: 10px;
+  padding: 5px;
+  height: auto;
+  gap: 10px;
+  background-color: var(--color-2);
+  min-height: 60px;
+}
+
+.subGroupBtn {
+  border-radius: 10px;
+  width: 100%;
+  height: 30px;
+  cursor: pointer;
+  background-color: var(--color-1);
+  border: 2px solid var(--color-5);
+}
